Refactor docker auth server
